requirements.txt
selene==2.0.0rc4
pytest==7.4.2
allure-pytest
